requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-taggit==1.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
idna==2.10
Pillow==8.1.2
pynput==1.7.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
